feat : Spring Security 설정 및 인가/인증 설정 구현
1. Spring Security 설정 및 인가/인증 설정 구현
- Spring Security 설정
- HTTP 및 그외 에러시 에러 메시지 일부 구현
- 사용자 정보 수정 및 탈퇴시 accessToken 사용하도록 변경
- 로그인후 받은 accessToken을 Header에 Authentication : Bearer {accessToken} 넣으면 api 요청 가능함
- 사용자 정보 추가 및 수정시 중복 이메일 사용 불가하도록 설정
- 사용자 정보 수정시 이메일은 수정 불가능하도록 업데이트 사항에서 제외

+++ src/main/java/com/project/mog/service/users/UsersService.java
@@ -1,11 +1,13 @@
 package com.project.mog.service.users;
 
+
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import com.project.mog.controller.login.LoginRequest;
@@ -39,6 +41,8 @@ public class UsersService {
 
 
 		public UsersDto createUser(UsersDto usersDto) {
+			UsersEntity isDuplicated = usersRepository.findByEmail(usersDto.getEmail());
+			if(isDuplicated!=null) throw new IllegalArgumentException("중복된 아이디입니다");
 			UsersEntity uEntity = usersRepository.save(usersDto.toEntity());
 			
 			return UsersDto.toDto(uEntity);
@@ -46,25 +50,34 @@ public class UsersService {
 
 
 		public Optional<UsersDto> getUser(Long usersId) {
-			
 			return usersRepository.findById(usersId).map(uEntity->UsersDto.toDto(uEntity));
 		}
 
 
-		public UsersDto deleteUser(Long usersId) {
-			UsersEntity usersEntity =usersRepository.findById(usersId).orElseThrow(()->new IllegalArgumentException(usersId+"가 존재하지 않습니다"));
+		public UsersDto deleteUser(Long usersId, String authEmail) {
+			UsersEntity currentUser = usersRepository.findByEmail(authEmail);
+			UsersEntity targetUser = usersRepository.findById(usersId).orElseThrow(()->new RuntimeException("삭제할 사용자를 찾을 수 없습니다"));
+			
+			//권한을 가진 유저의 수정 요청인지 확인
+			if(currentUser.getUsersId()!=targetUser.getUsersId()) throw new AccessDeniedException("자기 자신만 삭제 가능합니다");
 			
 			usersRepository.deleteById(usersId);
-			return UsersDto.toDto(usersEntity);
+			return UsersDto.toDto(targetUser);
 		}
 
 
-		public UsersDto editUser(UsersDto usersDto, Long usersId) {
+		public UsersDto editUser(UsersDto usersDto, Long usersId, String authEmail) {
+			UsersEntity currentUser = usersRepository.findByEmail(authEmail);
+			UsersEntity targetUser = usersRepository.findById(usersId).orElseThrow(()->new RuntimeException("수정할 사용자를 찾을 수 없습니다"));
+			
+			//권한을 가진 유저의 수정 요청인지 확인
+			if(currentUser.getUsersId()!=targetUser.getUsersId()) throw new AccessDeniedException("자기 자신만 수정 가능합니다");
+			
+			
 			UsersEntity usersEntity =usersRepository.findById(usersId).orElseThrow(()->new IllegalArgumentException(usersId+"가 존재하지 않습니다"));
 			BiosEntity biosEntity = biosRepository.findByUser(usersEntity);
 			AuthEntity authEntity = authRepository.findByUser(usersEntity);
 			usersEntity.setUsersName(usersDto.getUsersName());
-			usersEntity.setEmail(usersDto.getEmail());
 			usersEntity.setProfileImg(usersDto.getProfileImg()!=null?usersDto.getProfileImg():null);
 			usersEntity.setUpdateDate();
 			
@@ -86,6 +99,7 @@ public class UsersService {
 			
 			//authDto 업데이트
 			authEntity.setPassword(usersDto.getAuthDto().getPassword());
+			
 			return UsersDto.toDto(usersEntity);
 		}
